+++ tests/src/rules/prefer-default-export.js
@@ -7,6 +7,10 @@ import { version as tsEslintVersion } from 'typescript-eslint-parser/package.jso
 const ruleTester = new RuleTester();
 const rule = require('../../../src/rules/prefer-default-export');
 
+const SINGLE_EXPORT_ERROR_MESSAGE = 'Prefer default export on a file with single export.';
+const ANY_EXPORT_ERROR_MESSAGE = 'Prefer default export to be present on every file that has export.';
+
+// test cases for default option { target: 'single' }
 ruleTester.run('prefer-default-export', rule, {
   valid: [].concat(
     test({
@@ -108,7 +112,7 @@ ruleTester.run('prefer-default-export', rule, {
         export function bar() {};`,
       errors: [{
         type: 'ExportNamedDeclaration',
-        message: 'Prefer default export.',
+        message: SINGLE_EXPORT_ERROR_MESSAGE,
       }],
     }),
     test({
@@ -116,7 +120,7 @@ ruleTester.run('prefer-default-export', rule, {
         export const foo = 'foo';`,
       errors: [{
         type: 'ExportNamedDeclaration',
-        message: 'Prefer default export.',
+        message: SINGLE_EXPORT_ERROR_MESSAGE,
       }],
     }),
     test({
@@ -125,7 +129,7 @@ ruleTester.run('prefer-default-export', rule, {
         export { foo };`,
       errors: [{
         type: 'ExportSpecifier',
-        message: 'Prefer default export.',
+        message: SINGLE_EXPORT_ERROR_MESSAGE,
       }],
     }),
     test({
@@ -133,7 +137,7 @@ ruleTester.run('prefer-default-export', rule, {
         export const { foo } = { foo: "bar" };`,
       errors: [{
         type: 'ExportNamedDeclaration',
-        message: 'Prefer default export.',
+        message: SINGLE_EXPORT_ERROR_MESSAGE,
       }],
     }),
     test({
@@ -141,7 +145,7 @@ ruleTester.run('prefer-default-export', rule, {
         export const { foo: { bar } } = { foo: { bar: "baz" } };`,
       errors: [{
         type: 'ExportNamedDeclaration',
-        message: 'Prefer default export.',
+        message: SINGLE_EXPORT_ERROR_MESSAGE,
       }],
     }),
     test({
@@ -149,10 +153,183 @@ ruleTester.run('prefer-default-export', rule, {
         export const [a] = ["foo"]`,
       errors: [{
         type: 'ExportNamedDeclaration',
-        message: 'Prefer default export.',
+        message: SINGLE_EXPORT_ERROR_MESSAGE,
       }],
     }),
   ],
+});
+
+// test cases for { target: 'any' }
+ruleTester.run('prefer-default-export', rule, {
+  // Any exporting file must contain default export
+  valid: [].concat(
+    test({
+      code: `
+          export default function bar() {};`,
+      options: [{
+        target: 'any',
+      }],
+    }),
+    test({
+      code: `
+              export const foo = 'foo';
+              export const bar = 'bar';
+              export default 42;`,
+      options: [{
+        target: 'any',
+      }],
+    }),
+    test({
+      code: `
+            export default a = 2;`,
+      options: [{
+        target: 'any',
+      }],
+    }),
+    test({
+      code: `
+            export const a = 2;
+            export default function foo() {};`,
+      options: [{
+        target: 'any',
+      }],
+    }),
+    test({
+      code: `
+          export const a = 5;
+          export function bar(){};
+          let foo;
+          export { foo as default }`,
+      options: [{
+        target: 'any',
+      }],
+    }),
+    test({
+      code: `
+          export * from './foo';`,
+      options: [{
+        target: 'any',
+      }],
+    }),
+    test({
+      code: `export Memory, { MemoryValue } from './Memory'`,
+      parser: parsers.BABEL_OLD,
+      options: [{
+        target: 'any',
+      }],
+    }),
+    // no exports at all
+    test({
+      code: `
+            import * as foo from './foo';`,
+      options: [{
+        target: 'any',
+      }],
+    }),
+    test({
+      code: `const a = 5;`,
+      options: [{
+        target: 'any',
+      }],
+    }),
+    // es2022: Arbitrary module namespae identifier names
+    testVersion('>= 8.7', () => ({
+      code: 'export const a = 4; let foo; export { foo as "default" };',
+      options: [{
+        target: 'any',
+      }],
+      parserOptions: { ecmaVersion: 2022 },
+    })),
+  ),
+  // { target: 'any' } invalid cases when any exporting file must contain default export but does not
+  invalid: [].concat(
+    test({
+      code: `
+        export const foo = 'foo';
+        export const bar = 'bar';`,
+      options: [{
+        target: 'any',
+      }],
+      errors: [{
+        message: ANY_EXPORT_ERROR_MESSAGE,
+      }],
+    }),
+    test({
+      code: `
+        export const foo = 'foo';
+        export function bar() {};`,
+      options: [{
+        target: 'any',
+      }],
+      errors: [{
+        message: ANY_EXPORT_ERROR_MESSAGE,
+      }],
+    }),
+    test({
+      code: `
+        let foo, bar;
+        export { foo, bar }`,
+      options: [{
+        target: 'any',
+      }],
+      errors: [{
+        message: ANY_EXPORT_ERROR_MESSAGE,
+      }],
+    }),
+    test({
+      code: `
+        let item;
+        export const foo = item;
+        export { item };`,
+      options: [{
+        target: 'any',
+      }],
+      errors: [{
+        message: ANY_EXPORT_ERROR_MESSAGE,
+      }],
+    }),
+    test({
+      code: 'export { a, b } from "foo.js"',
+      parser: parsers.BABEL_OLD,
+      options: [{
+        target: 'any',
+      }],
+      errors: [{
+        message: ANY_EXPORT_ERROR_MESSAGE,
+      }],
+    }),
+    test({
+      code: `
+        const foo = 'foo';
+        export { foo };`,
+      options: [{
+        target: 'any',
+      }],
+      errors: [{
+        message: ANY_EXPORT_ERROR_MESSAGE,
+      }],
+    }),
+    test({
+      code: `
+        export const { foo } = { foo: "bar" };`,
+      options: [{
+        target: 'any',
+      }],
+      errors: [{
+        message: ANY_EXPORT_ERROR_MESSAGE,
+      }],
+    }),
+    test({
+      code: `
+        export const { foo: { bar } } = { foo: { bar: "baz" } };`,
+      options: [{
+        target: 'any',
+      }],
+      errors: [{
+        message: ANY_EXPORT_ERROR_MESSAGE,
+      }],
+    }),
+  ),
 });
 
 context('TypeScript', function () {